refactor: one folder for the domain data modules
`@db/*` mapped to `src/app/*/db.ts` — a glob that reached into any sibling
folder for a file with the right name. Nine of the ten folders it addressed
held nothing else: six contained only `db.ts`, three only `db.ts` and its
test. The folders were addressing, not grouping.

They are now `src/app/db/<domain>.ts`, beside the `core/` client that was
already there, and the alias is a plain directory map. **No import specifier
changed** — all 99 `@db/*` imports read exactly as before.

The shape had hidden a duplicate: `src/app/profiles/db.ts` was a one-line
`export * from '../profile/db'`, so `@db/profile` and `@db/profiles` were two
names for one module — and the shim had 14 callers to the original's 1.
Invisible while they sat in different folders. Collapsed into
`src/app/db/profiles.ts`, keeping the name the callers already used.

It had also hidden a second spelling. Four route files imported
`@app/factions/db`, `@app/migrations/db` and `@app/publications/db`, reaching
the same modules through the other alias — the same "an alias is not a path"
shape that put holes in the kit's import rule twice this week. That spelling
no longer resolves, so every domain import now goes through `@db/*`.

`factions/` and `faq/` keep their real domain logic (`authoringState`,
`catalogue`, `faqEditingSession`) and are honest folders now rather than
wrappers around one file.

Two guards moved with it. `assert-no-convex-skip-in-domain-db.mjs` walked the
whole app tree hunting for files named `db.ts`, which was the only way to find
them before; it now scans `src/app/db`, and I verified it still fails on a
planted `'skip'`. The kit's import rule loses the `@app/*/db` and
`**/src/app/*/db` patterns it needed for the old layout — ten probes confirm
every boundary still blocks, and `import type`, `@game` and `Link` still pass.

Verified: typecheck, lint, knip, format, check:convex-skip, check:css-orphans,
451 app tests, 279 story tests, publisher:release:verify, and the full e2e
suite.
