+++ .oxlintrc.json
@@ -36,20 +36,15 @@
             "patterns": [
               {
                 "group": ["convex", "convex/*", "**/convex/**", "../**/convex/**"],
-                "message": "A component must not fetch. Convex belongs in a domain module (`src/app/<domain>/db.ts`); take the result as a prop."
+                "message": "A component must not fetch. Convex belongs in a domain module (`src/app/db/<domain>.ts`); take the result as a prop."
               },
               {
                 "group": [
                   "@db/**",
                   "@app/db",
                   "@app/db/**",
-                  "@app/*/db",
-                  "@app/*/db.*",
                   "**/src/app/db/**",
-                  "**/src/app/*/db",
-                  "**/src/app/*/db.*",
-                  "../**/db",
-                  "../**/db.*",
+                  "../**/db/*",
                   "../**/db/core",
                   "../**/db/core/*"
                 ],
@@ -106,7 +101,7 @@
                   "@ui/**",
                   "@db/**"
                 ],
-                "message": "The server must not import client code. A contract both sides need — a Zod schema, a tag vocabulary — belongs in `src/shared/`, which both may import. `@db/*` resolves to `src/app/<domain>/db.ts`, so it is client code too."
+                "message": "The server must not import client code. A contract both sides need — a Zod schema, a tag vocabulary — belongs in `src/shared/`, which both may import. `@db/*` resolves to `src/app/db/<domain>.ts`, so it is client code too."
               },
               {
                 "group": ["**/_generated/server"],
@@ -134,7 +129,7 @@
                   "@ui/**",
                   "@db/**"
                 ],
-                "message": "The publisher Worker is a separate deploy artifact and must not import application code. Shared runtime contracts belong in `src/shared/`. `@db/*` resolves to `src/app/<domain>/db.ts`, so it is application code too."
+                "message": "The publisher Worker is a separate deploy artifact and must not import application code. Shared runtime contracts belong in `src/shared/`. `@db/*` resolves to `src/app/db/<domain>.ts`, so it is application code too."
               },
               {
                 "group": ["**/_generated/server"],
