+++ docs/README.md
@@ -6,8 +6,8 @@ Quick reference for understanding and working with the codebase.
 
 **Starting a new feature?**
 1. Routes: `src/app/routes/` (file-based routing; `_app` is a pathless layout)
-2. Domain logic: `src/app/<domain>/db.ts` (loaders plus live query and mutation hooks)
-3. Schemas: `src/app/<domain>/validation.ts` and `src/game/schema/` (Zod schemas)
+2. Domain logic: `src/app/db/<domain>.ts` (loaders plus live query and mutation hooks)
+3. Schemas: `src/shared/<domain>/validation.ts` and `src/game/schema/` (Zod schemas)
 4. Validation standard: [`docs/data-layer.md`](./data-layer.md) (Convex `v` + shared Zod)
 5. UI: every published component lives in `src/app/ui/<category>` (alias `@ui/*`), and the component
    taxonomy in [`AGENTS.md`](../AGENTS.md#component-taxonomy) decides which category. Lint holds one
