--- src/app/profiles/db.ts
@@ -1,1 +1,0 @@
-export * from '../profile/db';
